encoder now in sud2sat
